+++ Assets/Scripts/Beam/DiffusionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6795f3f15873274448a121ea9abbdf99
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
